sls_native/requirements.txt
psycopg2-binary
python-dotenv
raven
werkzeug
SQLAlchemy